update fcm token on login
