+++ app/src/main/java/com/codefusiongroup/gradshub/authentication/login/LoginModel.java
@@ -1,15 +1,19 @@
 package com.codefusiongroup.gradshub.authentication.login;
 
+import android.content.Context;
 import android.util.Log;
 
 import androidx.annotation.VisibleForTesting;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.codefusiongroup.gradshub.authentication.AuthenticationAPI;
 import com.codefusiongroup.gradshub.common.GradsHubApplication;
+import com.codefusiongroup.gradshub.common.UserPreferences;
 import com.codefusiongroup.gradshub.common.network.ApiBaseResponse;
 import com.codefusiongroup.gradshub.common.network.ApiProvider;
 import com.codefusiongroup.gradshub.common.models.User;
 import com.codefusiongroup.gradshub.common.network.ApiResponseConstants;
+import com.codefusiongroup.gradshub.messaging.MessagingAPI;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 
@@ -27,6 +31,8 @@ public class LoginModel implements LoginContract.ILoginModel {
     private static LoginModel loginModel = null;
     private final LoginContract.ILoginPresenter mPresenter;
 
+    private AppCompatActivity mActivity;
+    private UserPreferences mUserPreferences;
 
     private LoginModel(LoginContract.ILoginPresenter presenter) {
         mPresenter = presenter;
@@ -67,6 +73,15 @@ public class LoginModel implements LoginContract.ILoginModel {
                         mPresenter.setLoginResponseCode(SUCCESS_CODE);
                         mPresenter.setLoginResponseMessage(jsonObject.get("message").getAsString());
                         mPresenter.setCurrentUser(user);
+
+                        mUserPreferences = UserPreferences.getInstance();
+                        Context ctx = GradsHubApplication.getContext();
+                        if(mUserPreferences.isTokenChanged(ctx)){
+                            // Update the FCM token
+                            String token = mUserPreferences.getFCMToken(ctx);
+                            updateUserToken(user.getUserID(),token);
+                        }
+
                     }
                     else {
                         ApiBaseResponse apiDefault = new Gson().fromJson(jsonObject, ApiBaseResponse.class);
@@ -84,12 +99,48 @@ public class LoginModel implements LoginContract.ILoginModel {
 
             }
 
+
             @Override
             public void onFailure(Call<JsonObject> call, Throwable t) {
                 Log.i(TAG, "onFailure() executed");
                 mPresenter.setLoginResponseCode(ApiResponseConstants.SERVER_FAILURE_CODE);
                 mPresenter.setLoginResponseMessage(ApiResponseConstants.SERVER_FAILURE_MSG);
                 mPresenter.onLoginRequestFinished();
+                t.printStackTrace();
+            }
+
+        });
+
+    }
+
+
+    private void updateUserToken(String userID, String token) {
+
+        HashMap<String, String> params = new HashMap<>();
+        params.put("user_id", userID);
+        params.put("fcm_token", token);
+
+        MessagingAPI messagingAPI = ApiProvider.getMessageApiService();
+
+        messagingAPI.updateUserFCMToken(params).enqueue(new Callback<JsonObject>() {
+
+            @Override
+            public void onResponse(Call<JsonObject> call, retrofit2.Response<JsonObject> response) {
+
+                if ( response.isSuccessful() ) {
+                    JsonObject jsonObject = response.body();
+//                    GradsHubApplication.showToast("successfully updated your token for messaging.");
+                    Log.i(TAG, " update user fcm token response.isSuccessful() = true");
+                }
+                else {
+                    Log.i(TAG, "response.isSuccessful() = false");
+                }
+
+            }
+
+            @Override
+            public void onFailure(Call<JsonObject> call, Throwable t) {
+                GradsHubApplication.showToast("your token has been refreshed, please refresh page to continue receiving messages.");
                 t.printStackTrace();
             }
