+++ app/src/main/java/com/codefusiongroup/gradshub/common/UserPreferences.java
@@ -17,6 +17,7 @@ public class UserPreferences {
     private final String LOGIN_STATE = "is_loggedIn";
     private final String PREF_NAME = "com.codefusiongroup.gradshub.PREFERENCE_FILE_KEY";
     private final String FCM_TOKEN = "fcm_token";
+    private final String FCM_TOKEN_CHANGED = "fcm_token_changed";
 
     private SharedPreferences.Editor editor;
     private static UserPreferences instance;
@@ -79,6 +80,7 @@ public class UserPreferences {
         editor.apply();
     }
 
+    // FCM TOKEN
     public void saveFCMToken(String fcmToken, Context context){
         editor = context.getSharedPreferences(PREF_NAME,Context.MODE_PRIVATE).edit();
         editor.putString(FCM_TOKEN,fcmToken);
@@ -86,6 +88,13 @@ public class UserPreferences {
     }
     public String getFCMToken(Context context){
         return context.getSharedPreferences(PREF_NAME,Context.MODE_PRIVATE).getString(FCM_TOKEN,"Token Not Found");
+    }
+    public void tokenChanged(Context context) {
+        editor = context.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE).edit().putBoolean(FCM_TOKEN_CHANGED, true);
+        editor.apply();
+    }
+    public boolean isTokenChanged(Context context) {
+        return context.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE).getBoolean(FCM_TOKEN_CHANGED, false);
     }
 
 }
