+++ app/src/main/java/com/codefusiongroup/gradshub/messaging/MessagingService.java
@@ -9,6 +9,7 @@ import android.content.Intent;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.NotificationCompat;
 
 import com.codefusiongroup.gradshub.R;
@@ -165,44 +166,13 @@ public class MessagingService extends FirebaseMessagingService implements ChatMe
         Log.i(TAG, "onNewToken() executed, new token is: " + token);
 
         // Save the user token for later usage
+        Context ctx = GradsHubApplication.getContext();
         mUserPreferences = UserPreferences.getInstance();
-        mUserPreferences.saveFCMToken(token,getApplicationContext());
+        mUserPreferences.saveFCMToken(token,ctx);
+        Log.i(TAG, "Shared preferences token " + mUserPreferences.getFCMToken(ctx));
+        mUserPreferences.tokenChanged(ctx);
 
         //updateUserToken(userID, token);
-    }
-
-
-    private void updateUserToken(String userID, String token) {
-
-        HashMap<String, String> params = new HashMap<>();
-        params.put("user_id", userID);
-        params.put("fcm_token", token);
-
-        MessagingAPI messagingAPI = ApiProvider.getMessageApiService();
-
-        messagingAPI.updateUserFCMToken(params).enqueue(new Callback<JsonObject>() {
-
-            @Override
-            public void onResponse(Call<JsonObject> call, retrofit2.Response<JsonObject> response) {
-
-                if ( response.isSuccessful() ) {
-                    JsonObject jsonObject = response.body();
-                    GradsHubApplication.showToast("successfully updated your token for messaging.");
-                }
-                else {
-                    Log.i(TAG, "response.isSuccessful() = false");
-                }
-
-            }
-
-            @Override
-            public void onFailure(Call<JsonObject> call, Throwable t) {
-                GradsHubApplication.showToast("your token has been refreshed, please refresh page to continue receiving messages.");
-                t.printStackTrace();
-            }
-
-        });
-
     }
